backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.0
django-filter==23.5
django-otp==1.2.0
django-ratelimit==4.1.0
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
resend==2.13.1
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1
whitenoise==6.6.0
